Remove vector class, make zero and inf global functions
